HadoopUtils and YarnUtils improvements (#1093)
* HadoopUtils and YarnUtils improvements
