+++ streamx-common/src/main/scala/com/streamxhub/streamx/common/util/HttpClientUtils.scala
@@ -27,7 +27,7 @@ import org.apache.http.client.utils.URIBuilder
 import org.apache.http.config.RegistryBuilder
 import org.apache.http.entity.StringEntity
 import org.apache.http.impl.auth.SPNegoSchemeFactory
-import org.apache.http.impl.client.{BasicCredentialsProvider, HttpClientBuilder, HttpClients}
+import org.apache.http.impl.client.{BasicCredentialsProvider, CloseableHttpClient, HttpClientBuilder, HttpClients}
 import org.apache.http.impl.conn.PoolingHttpClientConnectionManager
 import org.apache.http.message.BasicNameValuePair
 import org.apache.http.util.EntityUtils
@@ -35,10 +35,8 @@ import org.apache.http.util.EntityUtils
 import java.io.UnsupportedEncodingException
 import java.net.URISyntaxException
 import java.nio.charset.{Charset, StandardCharsets}
-import java.security.{Principal, PrivilegedExceptionAction}
+import java.security.Principal
 import java.util
-import javax.security.auth.Subject
-import javax.security.auth.login.{Configuration, LoginContext}
 import scala.collection.JavaConversions._
 
 object HttpClientUtils {
@@ -58,23 +56,6 @@ object HttpClientUtils {
    * @return
    */
   private[this] def getHttpClient = HttpClients.custom.setConnectionManager(connectionManager).build
-
-  private[this] def getSpengoHttpClient = {
-    val builder = HttpClientBuilder.create()
-    val authSchemeRegistry = RegistryBuilder.create[AuthSchemeProvider]
-      .register(AuthSchemes.SPNEGO, new SPNegoSchemeFactory(true)).build
-    builder.setDefaultAuthSchemeRegistry(authSchemeRegistry)
-
-    val credentialsProvider = new BasicCredentialsProvider
-    credentialsProvider.setCredentials(new AuthScope(null, -1, null), new Credentials {
-      override def getUserPrincipal: Principal = null
-
-      override def getPassword: String = null
-    })
-    builder.setDefaultCredentialsProvider(credentialsProvider)
-    builder.setConnectionManager(connectionManager).build()
-  }
-
 
   private[this] def getHttpGet(url: String, params: util.Map[String, AnyRef] = null, config: RequestConfig = null): HttpGet = {
     val httpGet = params match {
@@ -186,34 +167,59 @@ object HttpClientUtils {
    * @return
    */
   private[this] def getResult(request: HttpRequestBase): String = {
-
-    Subject.doAs(getSubject, new PrivilegedExceptionAction[String] {
-      override def run(): String = {
-        val httpClient = getSpengoHttpClient
-        try {
-          val response = httpClient.execute(request)
-          val entity = response.getEntity
-          if (entity != null) { // long len = entity.getContentLength();// -1 表示长度未知
-            val result = EntityUtils.toString(entity)
-            response.close()
-            result
-          } else null
-        } catch {
-          case e: Exception => throw e
-        }
-      }
-    })
-  }
-
-  private[this] def getSubject: Subject = {
-    val enableLoginConfig = System.getProperty("java.security.auth.login.config")
-    if (null != enableLoginConfig) {
-      val loginConfig = Configuration.getConfiguration;
-      val lc = new LoginContext("httpClient", null, null, loginConfig)
-      lc.login()
-      return lc.getSubject
+    val httpClient = getHttpClient
+    try {
+      val response = httpClient.execute(request)
+      val entity = response.getEntity
+      if (entity != null) { // long len = entity.getContentLength();// -1 表示长度未知
+        val result = EntityUtils.toString(entity)
+        response.close()
+        result
+      } else null
+    } catch {
+      case e: Exception => throw e
     }
-    new Subject()
   }
+
+  /**
+   * @param url
+   * @return
+   */
+  def httpAuthGetRequest(url: String, config: RequestConfig): String = {
+
+    def getHttpAuthClient: CloseableHttpClient = {
+      val credentialsProvider = new BasicCredentialsProvider
+      credentialsProvider.setCredentials(
+        new AuthScope(null, -1, null),
+        new Credentials {
+          override def getUserPrincipal: Principal = null
+
+          override def getPassword: String = null
+        })
+
+      val authSchemeRegistry = RegistryBuilder.create[AuthSchemeProvider]
+        .register(AuthSchemes.SPNEGO, new SPNegoSchemeFactory(true)).build
+
+      HttpClientBuilder.create()
+        .setDefaultAuthSchemeRegistry(authSchemeRegistry)
+        .setDefaultCredentialsProvider(credentialsProvider)
+        .setConnectionManager(connectionManager).build()
+    }
+
+    try {
+      val request = getHttpGet(url, null, config)
+      val response = getHttpAuthClient.execute(request)
+      val entity = response.getEntity
+      if (entity != null) { // long len = entity.getContentLength();// -1 表示长度未知
+        val result = EntityUtils.toString(entity)
+        response.close()
+        result
+      } else null
+    } catch {
+      case e: Exception => throw e
+    }
+
+  }
+
 
 }
